size type (#819)
* modify test

* add assert

* bump conan

* build failure

* build fail

* revert change

* merge

* conan

* blk alloc verification

* incrementing both size and cnt before checking

* review comments

* decrement vol and home blks ref cnt when all ios are completed

* add assert in blk allocatpr

* review comments

* review comments

* change default state

* remove preparing

* log dev superblock allocation size

* build failure

* check for unopened log store IDs

* call multi completions for error

* watchdog cp

* testing

* log messages

* review comments

* cp timer

* review comments

* review comments

* review comments

* size type

* merge

+++ src/engine/device/physical_dev.cpp
@@ -364,7 +364,7 @@ ssize_t PhysicalDev::sync_writev(const iovec* iov, int iovcnt, uint32_t size, ui
     }
 }
 
-void PhysicalDev::write_zero(uint32_t size, uint64_t offset, uint8_t* cookie) {
+void PhysicalDev::write_zero(uint64_t size, uint64_t offset, uint8_t* cookie) {
     drive_iface->write_zero(m_iodev.get(), size, offset, cookie);
 }
